+++ Assets/Scripts/Input/PlayerInputActions.cs
@@ -145,6 +145,15 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
                     ""processors"": """",
                     ""interactions"": """",
                     ""initialStateCheck"": false
+                },
+                {
+                    ""name"": ""ExitToMenu"",
+                    ""type"": ""Button"",
+                    ""id"": ""52af9b4a-a514-4093-b540-7f3354fd0fd1"",
+                    ""expectedControlType"": """",
+                    ""processors"": """",
+                    ""interactions"": """",
+                    ""initialStateCheck"": false
                 }
             ],
             ""bindings"": [
@@ -257,6 +266,17 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
                     ""action"": ""Spray"",
                     ""isComposite"": false,
                     ""isPartOfComposite"": false
+                },
+                {
+                    ""name"": """",
+                    ""id"": ""6cd64f73-76d6-4b3e-a929-309f225e5222"",
+                    ""path"": ""<Keyboard>/escape"",
+                    ""interactions"": ""Hold(duration=1.5)"",
+                    ""processors"": """",
+                    ""groups"": """",
+                    ""action"": ""ExitToMenu"",
+                    ""isComposite"": false,
+                    ""isPartOfComposite"": false
                 }
             ]
         }
@@ -271,6 +291,7 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
         m_Player_Reset = m_Player.FindAction("Reset", throwIfNotFound: true);
         m_Player_ToggleDebug = m_Player.FindAction("ToggleDebug", throwIfNotFound: true);
         m_Player_Spray = m_Player.FindAction("Spray", throwIfNotFound: true);
+        m_Player_ExitToMenu = m_Player.FindAction("ExitToMenu", throwIfNotFound: true);
     }
 
     ~@PlayerInputActions()
@@ -357,6 +378,7 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
     private readonly InputAction m_Player_Reset;
     private readonly InputAction m_Player_ToggleDebug;
     private readonly InputAction m_Player_Spray;
+    private readonly InputAction m_Player_ExitToMenu;
     /// <summary>
     /// Provides access to input actions defined in input action map "Player".
     /// </summary>
@@ -392,6 +414,10 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
         /// Provides access to the underlying input action "Player/Spray".
         /// </summary>
         public InputAction @Spray => m_Wrapper.m_Player_Spray;
+        /// <summary>
+        /// Provides access to the underlying input action "Player/ExitToMenu".
+        /// </summary>
+        public InputAction @ExitToMenu => m_Wrapper.m_Player_ExitToMenu;
         /// <summary>
         /// Provides access to the underlying input action map instance.
         /// </summary>
@@ -436,6 +462,9 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
             @Spray.started += instance.OnSpray;
             @Spray.performed += instance.OnSpray;
             @Spray.canceled += instance.OnSpray;
+            @ExitToMenu.started += instance.OnExitToMenu;
+            @ExitToMenu.performed += instance.OnExitToMenu;
+            @ExitToMenu.canceled += instance.OnExitToMenu;
         }
 
         /// <summary>
@@ -465,6 +494,9 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
             @Spray.started -= instance.OnSpray;
             @Spray.performed -= instance.OnSpray;
             @Spray.canceled -= instance.OnSpray;
+            @ExitToMenu.started -= instance.OnExitToMenu;
+            @ExitToMenu.performed -= instance.OnExitToMenu;
+            @ExitToMenu.canceled -= instance.OnExitToMenu;
         }
 
         /// <summary>
@@ -547,5 +579,12 @@ public partial class @PlayerInputActions: IInputActionCollection2, IDisposable
         /// <seealso cref="UnityEngine.InputSystem.InputAction.performed" />
         /// <seealso cref="UnityEngine.InputSystem.InputAction.canceled" />
         void OnSpray(InputAction.CallbackContext context);
+        /// <summary>
+        /// Method invoked when associated input action "ExitToMenu" is either <see cref="UnityEngine.InputSystem.InputAction.started" />, <see cref="UnityEngine.InputSystem.InputAction.performed" /> or <see cref="UnityEngine.InputSystem.InputAction.canceled" />.
+        /// </summary>
+        /// <seealso cref="UnityEngine.InputSystem.InputAction.started" />
+        /// <seealso cref="UnityEngine.InputSystem.InputAction.performed" />
+        /// <seealso cref="UnityEngine.InputSystem.InputAction.canceled" />
+        void OnExitToMenu(InputAction.CallbackContext context);
     }
 }